post_service/requirements.txt
grpcio==1.40.0
grpcio-tools==1.40.0
protobuf==3.17.3
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pytest==6.2.5
